backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.6.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
redis==5.0.1
python-dotenv==1.0.0
httpx==0.26.0
pandas==2.1.3
python-multipart==0.0.6
sse-starlette==1.8.2
requests==2.31.0
psycopg2-binary==2.9.9
openai==1.12.0 